# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "description": "Your API description",
        "contact": {
            "email": "your-email@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/expenses": {
            "get": {
                "summary": "Get a list of expenses",
                "operationId": "36e62ecfbcfde37686ea63fe89d0c0c0",
                "responses": {
                    "200": {
                        "description": "List of expenses"
                    }
                }
            },
            "post": {
                "summary": "Create a new expense",
                "operationId": "927e29ac31ae983af0a55ce5d487d554",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "property1": {
                                        "type": "type1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Expense created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/expenses/{id}": {
            "get": {
                "summary": "Get details of a specific expense",
                "operationId": "b8d5c673ecdbb6a51bea7ec7cc5d0e30",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the expense",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expense details"
                    },
                    "404": {
                        "description": "Expense not found"
                    }
                }
            },
            "put": {
                "summary": "Update a specific expense",
                "operationId": "36a4200b139cad8582ff017bbdb5750e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the expense",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "property1": {
                                        "type": "type1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Expense updated"
                    },
                    "404": {
                        "description": "Expense not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "summary": "Delete a specific expense",
                "operationId": "b261a1443844b958952ae83b03fbbfa3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the expense",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Expense deleted"
                    },
                    "404": {
                        "description": "Expense not found"
                    }
                }
            }
        },
        "/incomes": {
            "get": {
                "summary": "Get all incomes",
                "operationId": "68c947f4f4b1407943df0f86a84582ce",
                "responses": {
                    "200": {
                        "description": "List of incomes"
                    }
                }
            },
            "post": {
                "summary": "Create a new income",
                "operationId": "c036d918866bb5db62083a4a050c80d6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Income created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/incomes/{id}": {
            "get": {
                "summary": "Get a specific income",
                "operationId": "3f7bdae20e436918dfc171c484e2aef9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the income",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Income details"
                    },
                    "404": {
                        "description": "Income not found"
                    }
                }
            },
            "put": {
                "summary": "Update a specific income",
                "operationId": "4bff7a1416746a189298bb364d24ed78",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the income",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Income updated"
                    },
                    "404": {
                        "description": "Income not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "summary": "Delete a specific income",
                "operationId": "c6fe245d90f38f582a7a45fa9d6de21c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the income",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Income deleted"
                    },
                    "404": {
                        "description": "Income not found"
                    }
                }
            }
        },
...
    }
}